requirements.txt
# Core backend
fastapi==0.104.1        # Web framework
uvicorn[standard]==0.24.0  # ASGI server

# Database
sqlalchemy==2.0.23      # ORM

# Templates & forms
jinja2==3.1.2           # HTML templates
python-multipart==0.0.6 # Form & file uploads

# Authentication & sessions
authlib==1.3.0          # Google OAuth
httpx==0.27.0           # Required by authlib
itsdangerous==2.2.0     # SessionMiddleware dependency

# Environment variables
python-dotenv==1.0.0    # Load .env locally